+++ src/domain/models/__init__.py
@@ -2,12 +2,10 @@
 
 from src.domain.models.chat_message import ChatMessage
 from src.domain.models.message_role import MessageRole
-from src.domain.models.timestamp import Timestamp
 from src.domain.models.user import User
 
 __all__ = [
     "ChatMessage",
     "MessageRole",
-    "Timestamp",
     "User",
 ]